split the operations module
